Db update -> cardBalance, Account - balance, Transaction - receipt and payment from double to float + now apllying converted currency values when making deposits.

+++ src/Data/AspNetCoreTemplate.Data.Models/Account.cs
@@ -51,7 +51,7 @@
         public Currency Currency { get; set; }
 
         [Required]
-        public double AccountBalance { get; set; }
+        public decimal AccountBalance { get; set; }
 
         public string IBAN { get; set; }
     }
